third_party/WebKit/LayoutTests/fast/dom/Window/window-postmessage-user-gesture-expected.txt

Test that user gesture is kept when postMessage() to self or across frames.
Click me

PASS Stack postMessages on self window and the user gesture is only consumed once
PASS Bounce postMessages between self and frame window and the user gesture is only consumed once in self window
PASS Bounce postMessages between self and frame window and the user gesture is only consumed once in frame
PASS successfullyParsed is true

TEST COMPLETE
